Module notify_rust::server

Experimental server taking the place of your Desktop Environments Notification Server.

This is not nearly meant for anything but testing, as it only prints notifications to stdout. It does not respond properly either yet.

This server will not replace an already running notification server.

Structs

NotificationServer
An experimental notification server. See the module level documentation for more.

Constants

VERSION
Version of the crate equals the version server.