Module notify_rust::hints

NotificationHints allow you to pass extra information to the server.

Many of these are standardized by either:

http://www.galago-project.org/specs/notification/0.9/x344.html

https://developer.gnome.org/notification-spec/#hints

Which of these are actually implemented depends strongly on the Notification server you talk to. Usually the get_capabilities() gives some clues, but the standards usually mention much more than is actually available.

Enums

NotificationHint
All currently implemented NotificationHints that can be sent.

Constants

ACTION_ICONS
"action-icons"

CATEGORY
"category"

DESKTOP_ENTRY
"desktop-entry"

IMAGE_DATA
"image-data" if spec_version > 1.1;

IMAGE_DATA_1_0
"image-data" if spec_version < 1.1;

IMAGE_DATA_1_1
"image_data" if spec_version == 1.1

IMAGE_PATH
"image-path"

RESIDENT
"resident"

SOUND_FILE
"sound-file"

SOUND_NAME
"sound-name"

SUPPRESS_SOUND
"suppress-sound"

TRANSIENT
"transient"

URGENCY
"urgency"

X
"x"

Y
"y"

Functions

hint_from_key_val
convenience converting a name and value into a hint

image_spec
matching image data key for each spec version