Module notify_rust::hints

NotificationHints allow to pass extra information to the server.

Many of these are standardized by either:

http://www.galago-project.org/specs/notification/0.9/x344.html

https://developer.gnome.org/notification-spec/#hints

Which of these are actually implemented depends strongly on the Notification server you talk to. Usually the get_capabilities() gives some clues, but the standards usually mention much more than is actually available.

Enums

NotificationHint
All currently implemented NotificationHints that can be send.