Enum notify_rust::hints::NotificationHint

pub enum NotificationHint {
    ActionIcons(bool),
    Category(String),
    DesktopEntry(String),
    ImagePath(String),
    Resident(bool),
    SoundFile(String),
    SoundName(String),
    SuppressSound(bool),
    Transient(bool),
    X(i32),
    Y(i32),
    Urgency(NotificationUrgency),
    Custom(String, String),
    Invalid,
}

All currently implemented NotificationHints that can be send.

Variants

ActionIcons

If true, server may interpret action identifiers as named icons and display those.

Category

Check out:

http://www.galago-project.org/specs/notification/0.9/x211.html

https://developer.gnome.org/notification-spec/#categories

DesktopEntry

Name of the DesktopEntry representing the calling application. In case of "firefox.desktop" use "firefox". May be used to retrieve the correct icon.

ImagePath

Display the image at this path.

Resident

This does not work on all servers, however timeout=0 will do the job

SoundFile

Play the sound at this path.

SoundName

A themeable named sound from the freedesktop.org sound naming specification to play when the notification pops up. Similar to icon-name, only for sounds. An example would be "message-new-instant".

SuppressSound

Transient

When set the server will treat the notification as transient and by-pass the server's persistence capability, if it should exist. When set the server will treat the notification as transient and by-pass the server's persistence capability, if it should exist.

X

Y

Urgency

Pass me a NotificationUrgency, either Low, Normal or Critical

Custom

If you want to pass something entirely different.

Invalid

Only used by this NotificationServer implementation

Methods

impl NotificationHint

fn as_bool(&self) -> Option<bool>

fn as_i32(&self) -> Option<i32>

fn as_str(&self) -> Option<&str>