Struct notify_rust::NotificationHandle

pub struct NotificationHandle { /* fields omitted */ }

A handle to a shown notification.

This keeps a connection alive to ensure actions work on certain desktops.

Methods

impl NotificationHandle

pub fn wait_for_action<F>(self, invokation_closure: F) where
F: FnOnce(&str),

Waits for the user to act on a notification and then calls invokation_closure with the name of the corresponding action.

pub fn close(self)

Manually close the notification

pub fn on_close<F>(self, closure: F) where
F: FnOnce(),

Executes a closure after the notification has closed.

Example

Notification::new().summary("Time is running out")
                   .body("This will go away.")
                   .icon("clock")
                   .show()
                   .unwrap()
                   .on_close(|| println!("closed"));

pub fn update(&mut self)

Replace the original notification with an updated version

Example

let mut notification = Notification::new().summary("Latest News")
                                          .body("Bayern Dortmund 3:2")
                                          .show()
                                          .unwrap();

std::thread::sleep_ms(1_500);

notification.summary("Latest News (Correction)")
            .body("Bayern Dortmund 3:3");

notification.update();

Watch out for different implementations of the notification server! On plasma5 or instance, you should also change the appname, so the old message is really replaced and not just amended. Xfce behaves well, all others have not been tested by the developer.

pub fn id(&self) -> u32

Returns the Handle's id.